refactor(native-rd): extract VideoRecorder from CaptureVideoScreen
Move camera/preview/record/flip UI and its state machine into a child
component. Parent retains save responsibility and delegates back-button
exit via an imperative requestExit() handle. No user-facing behavior
change; the existing test suite passes unchanged.
