+++ apps/native-rd/src/components/VideoRecorder/index.ts
@@ -1,0 +1,2 @@
+export { VideoRecorder } from "./VideoRecorder";
+export type { VideoRecorderHandle, VideoRecorderProps } from "./VideoRecorder";
